Initial commit
Products sold on ecommerce platforms receive a huge number of ratings and reviews on a day-to-day basis. These features are based on user experience and shows how satisfied customers are with their purchases.  This project aims to help Flipkart better understand how their products are received by users. By extracting and analyzing the reviews, Flipkart can identify product issues and areas of improvement.